pkgdown/google0acd00d8574ec2df.html

Create index annotations at variable line positions in many sequences data (visualise_many_sequences() helper)

This function is called by visualise_many_sequences() to create the x/y position data for placing the index annotations on the graph. Its arguments are either intermediate variables produced by visualise_many_sequences(), or arguments of visualise_many_sequences() directly passed through.

Returns a dataframe with x_position, y_position, annotation, and type columns. type is always "Number" (unused, but for consistency with convert_sequences_to_annotations()). annotation is always the position of the base along the line. In this function, the count is reset each line (compared to counting consistenly along in convert_sequences_to_annotations()) because each line is a different sequence.

Usage

create_many_sequence_index_annotations(
  new_sequences_vector,
  original_sequences_vector,
  original_indices_to_annotate,
  annotation_interval,
  annotate_full_lines = FALSE,
  annotations_above = TRUE,
  annotation_vertical_position = 1/3
)

Arguments

new_sequences_vector

vector. The output of insert_at_indices() when used with identical arguments.

original_sequences_vector

vector. The vector of sequences used for plotting, that was originally given to visualise_many_sequences(). Must also have been used as input to insert_at_indices() to create new_sequences_vector.

original_indices_to_annotate

positive integer vector. The vector of lines (i.e. indices) of original_vector to be annotated. Read from index_annotation_lines argument to [visualise_many_sequences()]. Must also have been used as input to insert_at_indices() to create new_sequences_vector.

annotate_full_lines

logical. Whether annotations should be calculated up to the end of the longest line (TRUE) or to the end of each line being annotated (FALSE, default).

annotations_above

logical. Whether annotations should be drawn above (TRUE, default) or below (FALSE) each annotated line. Must also have been used as input to insert_at_indices() to create new_sequences_vector.

annotation_vertical_position

numeric. The vertical position above/below each annotated line that annotations should be drawn. Must also have been used as input to insert_at_indices() to create new_sequences_vector.

annnotation_interval

integer. The frequency at which numbers should be placed underneath indicating base index, starting counting from the leftmost base. Setting to 0 causes this function to return an empty dataframe.

Value

dataframe. A dataframe with columns x_position, y_position, annotation, and type, with one observation per annotation number that needs to be drawn onto the ggplot.

Examples

## Set up arguments (e.g. from visualise_many_sequences() call)
sequences_data <- example_many_sequences
index_annotation_interval <- 10
index_annotations_above <- TRUE
index_annotation_full_line <- FALSE
index_annotation_vertical_position <- 1/3

## Create sequences vector
sequences <- extract_and_sort_sequences(
    example_many_sequences,
    grouping_levels = c("family" = 8, "individual" = 2)
)

## Insert blank rows as needed
new_sequences <- insert_at_indices(
    sequences,
    insertion_indices = c(1, 23, 37),
    insert_before = index_annotations_above,
    insert = "",
    vert = index_annotation_vertical_position
)

## Create annnotation dataframe
create_many_sequence_index_annotations(
    new_sequences_vector = new_sequences,
    original_sequences_vector = sequences,
    annotation_interval = 10,
    annotate_full_lines = index_annotation_full_line,
    annotations_above = index_annotations_above,
    annotation_vertical_position = index_annotation_vertical_position
)
#> Error in create_many_sequence_index_annotations(new_sequences_vector = new_sequences,     original_sequences_vector = sequences, annotation_interval = 10,     annotate_full_lines = index_annotation_full_line, annotations_above = index_annotations_above,     annotation_vertical_position = index_annotation_vertical_position): argument "original_indices_to_annotate" is missing, with no default