Create dataframe of locations and rendered probabilities (visualise_methylation() helper)

This function takes the locations/probabilities/sequences input to visualise_methylation(), as well as the scaling and rounding to apply to the probability text, and produces a dataframe of the x and y coordinates to draw each probability at (i.e. inside the coloured box for each assessed base) and the probability text to draw inside each box.

Usage

rasterise_probabilities(
  modification_locations,
  modification_probabilities,
  sequences,
  sequence_text_scaling = c(-0.5, 256),
  sequence_text_rounding = 2
)

Arguments

modification_locations

character vector. One character value for each sequence, storing a condensed string (e.g. "3,6,9,12", produced via vector_to_string()) of the indices along the read at which modification was assessed. Indexing starts at 1.

modification_probabilities

character vector. One character value for each sequence, storing a condensed string (e.g. "0,128,255,15", produced via vector_to_string()) of the probability of methylation/modification at each assessed base.

Assumed to be Nanopore > SAM style modification stored as an 8-bit integer from 0 to 255, but changing other arguments could make this work on other scales.

sequences

character vector. One character value for each sequence, storing the actual DNA sequence.

sequence_text_scaling

numeric vector, length 2. The min and max possible probability values, used to facilitate scaling of the text in each to 0-1. Scaling is implemented as \(\frac{p - min}{max}\), so custom scalings (e.g. scaled to 0-9 space) can be implemented by setting this values as required.

Set to c(0, 1) to not scale at all i.e. print the raw integer probability values. It is recommended to also set sequence_text_rounding = 0 to print integers as the default value of 2 will result in e.g. "128.00".

Set to c(-0.5, 256) (default, results in \(\frac{p+0.5}{256}\)) to scale to the centre of the probability spaces defined by the SAMtools spec, where integer \(p\) represents the probability space from \(\frac{p}{256}\) to \(\frac{p+1}{256}\). This is slightly better at representing the uncertainty compared to c(0, 255) as strictly speaking 0 represents the probability space from 0.000 to 0.004 and 255 represents the probability space from 0.996 to 1.000, so scaling them to 0.002 and 0.998 respectively is a more accurate representation of the probability space they each represent. Setting c(0, 255) would scale such that 0 is exactly 0.000 and 255 is exactly 1.000, which is not as accurate so it discouraged.

sequence_text_rounding

integer. How many decimal places the text drawn in the boxes should be rounded to (defaults to 2). Ignored if sequence_text_type is "sequence" or "none".

Value

dataframe. Dataframe of x, y, and value (i.e. probability to draw).

Examples

d <- extract_and_sort_methylation(example_many_sequences)

## Unscaled i.e. integers
rasterise_probabilities(
    d$locations,
    d$probabilities,
    d$sequences,
    sequence_text_scaling = c(0, 1),
    sequence_text_rounding = 0
)
#>              x           y value
#> 1   0.02450980 0.990196078    29
#> 2   0.05392157 0.990196078   159
#> 3   0.08333333 0.990196078   155
#> 4   0.11274510 0.990196078   159
#> 5   0.14215686 0.990196078   220
#> 6   0.17156863 0.990196078   163
#> 7   0.20098039 0.990196078     2
#> 8   0.23039216 0.990196078    59
#> 9   0.25980392 0.990196078   170
#> 10  0.34803922 0.990196078   131
#> 11  0.37745098 0.990196078   177
#> 12  0.40686275 0.990196078   139
#> 13  0.49509804 0.990196078    72
#> 14  0.52450980 0.990196078   235
#> 15  0.55392157 0.990196078    75
#> 16  0.64215686 0.990196078   214
#> 17  0.67156863 0.990196078    73
#> 18  0.70098039 0.990196078    68
#> 19  0.78921569 0.990196078    48
#> 20  0.81862745 0.990196078    59
#> 21  0.84803922 0.990196078    81
#> 22  0.93627451 0.990196078    77
#> 23  0.96568627 0.990196078    41
#> 24  0.02450980 0.970588235   170
#> 25  0.05392157 0.970588235   236
#> 26  0.08333333 0.970588235   120
#> 27  0.11274510 0.970588235    36
#> 28  0.14215686 0.970588235   139
#> 29  0.17156863 0.970588235    50
#> 30  0.25980392 0.970588235   229
#> 31  0.28921569 0.970588235    99
#> 32  0.31862745 0.970588235    79
#> 33  0.40686275 0.970588235    41
#> 34  0.43627451 0.970588235   229
#> 35  0.46568627 0.970588235    42
#> 36  0.55392157 0.970588235   230
#> 37  0.58333333 0.970588235    34
#> 38  0.61274510 0.970588235    34
#> 39  0.70098039 0.970588235    27
#> 40  0.73039216 0.970588235   130
#> 41  0.75980392 0.970588235    77
#> 42  0.84803922 0.970588235     7
#> 43  0.87745098 0.970588235    79
#> 44  0.02450980 0.950980392   206
#> 45  0.05392157 0.950980392   141
#> 46  0.08333333 0.950980392   165
#> 47  0.11274510 0.950980392    80
#> 48  0.14215686 0.950980392   159
#> 49  0.17156863 0.950980392    84
#> 50  0.20098039 0.950980392   128
#> 51  0.23039216 0.950980392   173
#> 52  0.25980392 0.950980392   124
#> 53  0.34803922 0.950980392    62
#> 54  0.37745098 0.950980392   195
#> 55  0.40686275 0.950980392    19
#> 56  0.49509804 0.950980392    79
#> 57  0.52450980 0.950980392   183
#> 58  0.55392157 0.950980392   129
#> 59  0.64215686 0.950980392    39
#> 60  0.67156863 0.950980392   129
#> 61  0.70098039 0.950980392   126
#> 62  0.78921569 0.950980392   192
#> 63  0.81862745 0.950980392    45
#> 64  0.02450980 0.931372549   216
#> 65  0.05392157 0.931372549   221
#> 66  0.08333333 0.931372549    11
#> 67  0.11274510 0.931372549    81
#> 68  0.14215686 0.931372549     4
#> 69  0.17156863 0.931372549    61
#> 70  0.20098039 0.931372549   180
#> 71  0.23039216 0.931372549    79
#> 72  0.25980392 0.931372549   130
#> 73  0.28921569 0.931372549    13
#> 74  0.31862745 0.931372549   144
#> 75  0.34803922 0.931372549    31
#> 76  0.43627451 0.931372549   228
#> 77  0.46568627 0.931372549     4
#> 78  0.49509804 0.931372549   200
#> 79  0.58333333 0.931372549    23
#> 80  0.61274510 0.931372549   132
#> 81  0.64215686 0.931372549    98
#> 82  0.73039216 0.931372549    18
#> 83  0.75980392 0.931372549    82
#> 84  0.02450980 0.911764706    10
#> 85  0.05392157 0.911764706    56
#> 86  0.08333333 0.911764706   207
#> 87  0.11274510 0.911764706   134
#> 88  0.14215686 0.911764706   233
#> 89  0.17156863 0.911764706   212
#> 90  0.20098039 0.911764706    12
#> 91  0.23039216 0.911764706   116
#> 92  0.25980392 0.911764706    68
#> 93  0.28921569 0.911764706    78
#> 94  0.31862745 0.911764706   129
#> 95  0.40686275 0.911764706    46
#> 96  0.43627451 0.911764706   194
#> 97  0.46568627 0.911764706    51
#> 98  0.55392157 0.911764706    66
#> 99  0.58333333 0.911764706   253
#> 100 0.02450980 0.852941176    31
#> 101 0.05392157 0.852941176    56
#> 102 0.08333333 0.852941176   233
#> 103 0.11274510 0.852941176   241
#> 104 0.14215686 0.852941176    71
#> 105 0.17156863 0.852941176    31
#> 106 0.20098039 0.852941176   203
#> 107 0.23039216 0.852941176   190
#> 108 0.25980392 0.852941176   234
#> 109 0.28921569 0.852941176   254
#> 110 0.31862745 0.852941176   240
#> 111 0.34803922 0.852941176   124
#> 112 0.46568627 0.852941176    72
#> 113 0.49509804 0.852941176    64
#> 114 0.61274510 0.852941176   128
#> 115 0.64215686 0.852941176   127
#> 116 0.02450980 0.833333333   189
#> 117 0.05392157 0.833333333     9
#> 118 0.08333333 0.833333333   144
#> 119 0.11274510 0.833333333    71
#> 120 0.14215686 0.833333333    52
#> 121 0.17156863 0.833333333    34
#> 122 0.20098039 0.833333333    83
#> 123 0.23039216 0.833333333    40
#> 124 0.25980392 0.833333333    33
#> 125 0.28921569 0.833333333   111
#> 126 0.40686275 0.833333333    10
#> 127 0.43627451 0.833333333   182
#> 128 0.55392157 0.833333333    26
#> 129 0.58333333 0.833333333   242
#> 130 0.02450980 0.774509804    81
#> 131 0.05392157 0.774509804   245
#> 132 0.08333333 0.774509804   162
#> 133 0.11274510 0.774509804    32
#> 134 0.14215686 0.774509804   108
#> 135 0.17156863 0.774509804   233
#> 136 0.20098039 0.774509804   119
#> 137 0.23039216 0.774509804   232
#> 138 0.25980392 0.774509804   152
#> 139 0.34803922 0.774509804   161
#> 140 0.37745098 0.774509804   222
#> 141 0.40686275 0.774509804   128
#> 142 0.49509804 0.774509804   251
#> 143 0.52450980 0.774509804    83
#> 144 0.55392157 0.774509804   123
#> 145 0.64215686 0.774509804    91
#> 146 0.67156863 0.774509804   160
#> 147 0.70098039 0.774509804   189
#> 148 0.78921569 0.774509804   144
#> 149 0.81862745 0.774509804   250
#> 150 0.02450980 0.754901961   149
#> 151 0.05392157 0.754901961   181
#> 152 0.08333333 0.754901961   109
#> 153 0.11274510 0.754901961    88
#> 154 0.14215686 0.754901961   194
#> 155 0.17156863 0.754901961   108
#> 156 0.20098039 0.754901961   143
#> 157 0.23039216 0.754901961    30
#> 158 0.25980392 0.754901961    77
#> 159 0.34803922 0.754901961   122
#> 160 0.37745098 0.754901961    88
#> 161 0.40686275 0.754901961   153
#> 162 0.49509804 0.754901961    19
#> 163 0.52450980 0.754901961   244
#> 164 0.55392157 0.754901961     6
#> 165 0.64215686 0.754901961   215
#> 166 0.67156863 0.754901961   161
#> 167 0.75980392 0.754901961    79
#> 168 0.78921569 0.754901961   189
#> 169 0.02450980 0.735294118   147
#> 170 0.05392157 0.735294118   112
#> 171 0.08333333 0.735294118    58
#> 172 0.11274510 0.735294118    21
#> 173 0.14215686 0.735294118   217
#> 174 0.17156863 0.735294118    60
#> 175 0.20098039 0.735294118   252
#> 176 0.23039216 0.735294118   153
#> 177 0.25980392 0.735294118   255
#> 178 0.28921569 0.735294118    96
#> 179 0.31862745 0.735294118   142
#> 180 0.34803922 0.735294118   110
#> 181 0.37745098 0.735294118   147
#> 182 0.40686275 0.735294118   110
#> 183 0.43627451 0.735294118    57
#> 184 0.52450980 0.735294118    22
#> 185 0.55392157 0.735294118   163
#> 186 0.58333333 0.735294118   110
#> 187 0.67156863 0.735294118    19
#> 188 0.70098039 0.735294118   205
#> 189 0.73039216 0.735294118    83
#> 190 0.75980392 0.735294118   193
#> 191 0.02450980 0.558823529   163
#> 192 0.05392157 0.558823529   253
#> 193 0.08333333 0.558823529    33
#> 194 0.11274510 0.558823529   225
#> 195 0.14215686 0.558823529   207
#> 196 0.17156863 0.558823529   210
#> 197 0.20098039 0.558823529   213
#> 198 0.23039216 0.558823529   187
#> 199 0.25980392 0.558823529   251
#> 200 0.28921569 0.558823529   163
#> 201 0.31862745 0.558823529   168
#> 202 0.34803922 0.558823529   135
#> 203 0.37745098 0.558823529    81
#> 204 0.40686275 0.558823529   196
#> 205 0.43627451 0.558823529   134
#> 206 0.46568627 0.558823529   187
#> 207 0.49509804 0.558823529    78
#> 208 0.52450980 0.558823529   103
#> 209 0.55392157 0.558823529    52
#> 210 0.58333333 0.558823529   251
#> 211 0.61274510 0.558823529   144
#> 212 0.64215686 0.558823529    71
#> 213 0.67156863 0.558823529    47
#> 214 0.70098039 0.558823529   193
#> 215 0.73039216 0.558823529   145
#> 216 0.75980392 0.558823529   238
#> 217 0.84803922 0.558823529   163
#> 218 0.87745098 0.558823529   179
#> 219 0.02450980 0.500000000   191
#> 220 0.05392157 0.500000000    91
#> 221 0.08333333 0.500000000   194
#> 222 0.11274510 0.500000000    96
#> 223 0.14215686 0.500000000   204
#> 224 0.17156863 0.500000000     7
#> 225 0.20098039 0.500000000   129
#> 226 0.23039216 0.500000000   209
#> 227 0.25980392 0.500000000   139
#> 228 0.28921569 0.500000000    68
#> 229 0.31862745 0.500000000    88
#> 230 0.34803922 0.500000000    94
#> 231 0.37745098 0.500000000   109
#> 232 0.40686275 0.500000000   234
#> 233 0.43627451 0.500000000   200
#> 234 0.46568627 0.500000000   188
#> 235 0.49509804 0.500000000    72
#> 236 0.52450980 0.500000000   116
#> 237 0.55392157 0.500000000    73
#> 238 0.58333333 0.500000000   178
#> 239 0.61274510 0.500000000   209
#> 240 0.64215686 0.500000000   167
#> 241 0.67156863 0.500000000   105
#> 242 0.70098039 0.500000000   243
#> 243 0.73039216 0.500000000    62
#> 244 0.81862745 0.500000000   155
#> 245 0.84803922 0.500000000   193
#> 246 0.02450980 0.480392157   176
#> 247 0.05392157 0.480392157   250
#> 248 0.08333333 0.480392157   122
#> 249 0.11274510 0.480392157   197
#> 250 0.14215686 0.480392157   146
#> 251 0.17156863 0.480392157   246
#> 252 0.20098039 0.480392157   203
#> 253 0.23039216 0.480392157   136
#> 254 0.25980392 0.480392157   152
#> 255 0.28921569 0.480392157    67
#> 256 0.31862745 0.480392157    71
#> 257 0.34803922 0.480392157    17
#> 258 0.37745098 0.480392157   144
#> 259 0.40686275 0.480392157    67
#> 260 0.43627451 0.480392157     1
#> 261 0.46568627 0.480392157   150
#> 262 0.49509804 0.480392157   133
#> 263 0.52450980 0.480392157   215
#> 264 0.55392157 0.480392157     8
#> 265 0.58333333 0.480392157   153
#> 266 0.61274510 0.480392157    68
#> 267 0.64215686 0.480392157    31
#> 268 0.67156863 0.480392157    26
#> 269 0.70098039 0.480392157   191
#> 270 0.78921569 0.480392157     4
#> 271 0.81862745 0.480392157    13
#> 272 0.02450980 0.460784314   122
#> 273 0.05392157 0.460784314   217
#> 274 0.08333333 0.460784314   108
#> 275 0.11274510 0.460784314     8
#> 276 0.14215686 0.460784314    66
#> 277 0.17156863 0.460784314    85
#> 278 0.20098039 0.460784314    34
#> 279 0.23039216 0.460784314   127
#> 280 0.25980392 0.460784314   205
#> 281 0.28921569 0.460784314    86
#> 282 0.31862745 0.460784314   130
#> 283 0.34803922 0.460784314   126
#> 284 0.37745098 0.460784314   203
#> 285 0.40686275 0.460784314   145
#> 286 0.43627451 0.460784314    27
#> 287 0.46568627 0.460784314   206
#> 288 0.49509804 0.460784314   145
#> 289 0.52450980 0.460784314    54
#> 290 0.55392157 0.460784314   191
#> 291 0.58333333 0.460784314    78
#> 292 0.61274510 0.460784314   125
#> 293 0.64215686 0.460784314   252
#> 294 0.67156863 0.460784314   108
#> 295 0.75980392 0.460784314    62
#> 296 0.78921569 0.460784314    55
#> 297 0.02450980 0.284313725   177
#> 298 0.05392157 0.284313725    29
#> 299 0.08333333 0.284313725   162
#> 300 0.17156863 0.284313725    79
#> 301 0.20098039 0.284313725    90
#> 302 0.28921569 0.284313725   250
#> 303 0.31862745 0.284313725   137
#> 304 0.40686275 0.284313725   113
#> 305 0.43627451 0.284313725   242
#> 306 0.52450980 0.284313725   115
#> 307 0.55392157 0.284313725    49
#> 308 0.64215686 0.284313725   253
#> 309 0.67156863 0.284313725   140
#> 310 0.75980392 0.284313725   196
#> 311 0.78921569 0.284313725   233
#> 312 0.87745098 0.284313725   174
#> 313 0.90686275 0.284313725   104
#> 314 0.02450980 0.264705882   104
#> 315 0.05392157 0.264705882    37
#> 316 0.08333333 0.264705882    50
#> 317 0.11274510 0.264705882    49
#> 318 0.14215686 0.264705882   104
#> 319 0.17156863 0.264705882    89
#> 320 0.20098039 0.264705882   213
#> 321 0.28921569 0.264705882    51
#> 322 0.31862745 0.264705882   220
#> 323 0.40686275 0.264705882   101
#> 324 0.43627451 0.264705882    39
#> 325 0.52450980 0.264705882    87
#> 326 0.55392157 0.264705882    94
#> 327 0.64215686 0.264705882   109
#> 328 0.67156863 0.264705882    48
#> 329 0.75980392 0.264705882   168
#> 330 0.78921569 0.264705882   235
#> 331 0.87745098 0.264705882   187
#> 332 0.90686275 0.264705882   225
#> 333 0.02450980 0.205882353   243
#> 334 0.05392157 0.205882353    50
#> 335 0.08333333 0.205882353   121
#> 336 0.11274510 0.205882353    98
#> 337 0.14215686 0.205882353    95
#> 338 0.17156863 0.205882353     7
#> 339 0.25980392 0.205882353   237
#> 340 0.28921569 0.205882353   105
#> 341 0.37745098 0.205882353   244
#> 342 0.40686275 0.205882353    69
#> 343 0.49509804 0.205882353   132
#> 344 0.52450980 0.205882353   249
#> 345 0.61274510 0.205882353    94
#> 346 0.64215686 0.205882353    79
#> 347 0.73039216 0.205882353     9
#> 348 0.75980392 0.205882353   170
#> 349 0.84803922 0.205882353   235
#> 350 0.87745098 0.205882353    11
#> 351 0.02450980 0.186274510    51
#> 352 0.05392157 0.186274510   190
#> 353 0.08333333 0.186274510    33
#> 354 0.11274510 0.186274510   181
#> 355 0.14215686 0.186274510   255
#> 356 0.17156863 0.186274510   241
#> 357 0.25980392 0.186274510   151
#> 358 0.28921569 0.186274510   186
#> 359 0.37745098 0.186274510   124
#> 360 0.40686275 0.186274510   196
#> 361 0.49509804 0.186274510     1
#> 362 0.52450980 0.186274510   142
#> 363 0.61274510 0.186274510   117
#> 364 0.64215686 0.186274510    84
#> 365 0.73039216 0.186274510   213
#> 366 0.75980392 0.186274510   249
#> 367 0.84803922 0.186274510   168
#> 368 0.02450980 0.166666667    60
#> 369 0.05392157 0.166666667   209
#> 370 0.08333333 0.166666667   185
#> 371 0.11274510 0.166666667   249
#> 372 0.14215686 0.166666667    68
#> 373 0.17156863 0.166666667   224
#> 374 0.20098039 0.166666667   124
#> 375 0.23039216 0.166666667    78
#> 376 0.25980392 0.166666667   101
#> 377 0.28921569 0.166666667   194
#> 378 0.37745098 0.166666667    26
#> 379 0.40686275 0.166666667   107
#> 380 0.49509804 0.166666667   168
#> 381 0.52450980 0.166666667    75
#> 382 0.61274510 0.166666667    53
#> 383 0.64215686 0.166666667     1
#> 384 0.73039216 0.166666667    27
#> 385 0.75980392 0.166666667    55
#> 386 0.78921569 0.166666667    29
#> 387 0.81862745 0.166666667   175
#> 388 0.02450980 0.107843137    49
#> 389 0.05392157 0.107843137   251
#> 390 0.08333333 0.107843137   241
#> 391 0.11274510 0.107843137   176
#> 392 0.14215686 0.107843137   189
#> 393 0.17156863 0.107843137   187
#> 394 0.20098039 0.107843137   166
#> 395 0.28921569 0.107843137    43
#> 396 0.31862745 0.107843137   235
#> 397 0.40686275 0.107843137   144
#> 398 0.43627451 0.107843137   137
#> 399 0.52450980 0.107843137     5
#> 400 0.55392157 0.107843137    93
#> 401 0.64215686 0.107843137   175
#> 402 0.67156863 0.107843137   106
#> 403 0.75980392 0.107843137   193
#> 404 0.78921569 0.107843137   198
#> 405 0.87745098 0.107843137   146
#> 406 0.90686275 0.107843137    48
#> 407 0.02450980 0.049019608   193
#> 408 0.05392157 0.049019608    24
#> 409 0.08333333 0.049019608   159
#> 410 0.11274510 0.049019608   106
#> 411 0.14215686 0.049019608   198
#> 412 0.17156863 0.049019608   206
#> 413 0.20098039 0.049019608   247
#> 414 0.23039216 0.049019608    55
#> 415 0.25980392 0.049019608   221
#> 416 0.28921569 0.049019608   106
#> 417 0.31862745 0.049019608   131
#> 418 0.40686275 0.049019608   198
#> 419 0.43627451 0.049019608    34
#> 420 0.52450980 0.049019608   105
#> 421 0.55392157 0.049019608   169
#> 422 0.64215686 0.049019608   231
#> 423 0.67156863 0.049019608    88
#> 424 0.75980392 0.049019608    27
#> 425 0.78921569 0.049019608   238
#> 426 0.87745098 0.049019608    51
#> 427 0.90686275 0.049019608    14
#> 428 0.02450980 0.029411765   161
#> 429 0.05392157 0.029411765   156
#> 430 0.08333333 0.029411765     9
#> 431 0.11274510 0.029411765    65
#> 432 0.14215686 0.029411765   198
#> 433 0.17156863 0.029411765   255
#> 434 0.20098039 0.029411765   245
#> 435 0.23039216 0.029411765   191
#> 436 0.31862745 0.029411765   174
#> 437 0.34803922 0.029411765    63
#> 438 0.43627451 0.029411765   155
#> 439 0.46568627 0.029411765   146
#> 440 0.55392157 0.029411765    13
#> 441 0.58333333 0.029411765    95
#> 442 0.67156863 0.029411765   228
#> 443 0.70098039 0.029411765   100
#> 444 0.78921569 0.029411765   132
#> 445 0.81862745 0.029411765    45
#> 446 0.84803922 0.029411765    49
#> 447 0.02450980 0.009803922   109
#> 448 0.05392157 0.009803922    86
#> 449 0.08333333 0.009803922    70
#> 450 0.11274510 0.009803922   169
#> 451 0.14215686 0.009803922   200
#> 452 0.17156863 0.009803922   112
#> 453 0.20098039 0.009803922   237
#> 454 0.23039216 0.009803922    69
#> 455 0.25980392 0.009803922   168
#> 456 0.28921569 0.009803922    97
#> 457 0.31862745 0.009803922   239
#> 458 0.34803922 0.009803922   188
#> 459 0.43627451 0.009803922   150
#> 460 0.46568627 0.009803922   208
#> 461 0.55392157 0.009803922   225
#> 462 0.58333333 0.009803922   190
#> 463 0.67156863 0.009803922   128
#> 464 0.70098039 0.009803922   252
#> 465 0.73039216 0.009803922   142
#> 466 0.75980392 0.009803922   224

## Scaled to 0-1, 3 dp
rasterize_probabilities(
    d$locations,
    d$probabilities,
    d$sequences,
    sequence_text_scaling = c(-0.5, 256),
    sequence_text_rounding = 3
)
#>              x           y value
#> 1   0.02450980 0.990196078 0.115
#> 2   0.05392157 0.990196078 0.623
#> 3   0.08333333 0.990196078 0.607
#> 4   0.11274510 0.990196078 0.623
#> 5   0.14215686 0.990196078 0.861
#> 6   0.17156863 0.990196078 0.639
#> 7   0.20098039 0.990196078 0.010
#> 8   0.23039216 0.990196078 0.232
#> 9   0.25980392 0.990196078 0.666
#> 10  0.34803922 0.990196078 0.514
#> 11  0.37745098 0.990196078 0.693
#> 12  0.40686275 0.990196078 0.545
#> 13  0.49509804 0.990196078 0.283
#> 14  0.52450980 0.990196078 0.920
#> 15  0.55392157 0.990196078 0.295
#> 16  0.64215686 0.990196078 0.838
#> 17  0.67156863 0.990196078 0.287
#> 18  0.70098039 0.990196078 0.268
#> 19  0.78921569 0.990196078 0.189
#> 20  0.81862745 0.990196078 0.232
#> 21  0.84803922 0.990196078 0.318
#> 22  0.93627451 0.990196078 0.303
#> 23  0.96568627 0.990196078 0.162
#> 24  0.02450980 0.970588235 0.666
#> 25  0.05392157 0.970588235 0.924
#> 26  0.08333333 0.970588235 0.471
#> 27  0.11274510 0.970588235 0.143
#> 28  0.14215686 0.970588235 0.545
#> 29  0.17156863 0.970588235 0.197
#> 30  0.25980392 0.970588235 0.896
#> 31  0.28921569 0.970588235 0.389
#> 32  0.31862745 0.970588235 0.311
#> 33  0.40686275 0.970588235 0.162
#> 34  0.43627451 0.970588235 0.896
#> 35  0.46568627 0.970588235 0.166
#> 36  0.55392157 0.970588235 0.900
#> 37  0.58333333 0.970588235 0.135
#> 38  0.61274510 0.970588235 0.135
#> 39  0.70098039 0.970588235 0.107
#> 40  0.73039216 0.970588235 0.510
#> 41  0.75980392 0.970588235 0.303
#> 42  0.84803922 0.970588235 0.029
#> 43  0.87745098 0.970588235 0.311
#> 44  0.02450980 0.950980392 0.807
#> 45  0.05392157 0.950980392 0.553
#> 46  0.08333333 0.950980392 0.646
#> 47  0.11274510 0.950980392 0.314
#> 48  0.14215686 0.950980392 0.623
#> 49  0.17156863 0.950980392 0.330
#> 50  0.20098039 0.950980392 0.502
#> 51  0.23039216 0.950980392 0.678
#> 52  0.25980392 0.950980392 0.486
#> 53  0.34803922 0.950980392 0.244
#> 54  0.37745098 0.950980392 0.764
#> 55  0.40686275 0.950980392 0.076
#> 56  0.49509804 0.950980392 0.311
#> 57  0.52450980 0.950980392 0.717
#> 58  0.55392157 0.950980392 0.506
#> 59  0.64215686 0.950980392 0.154
#> 60  0.67156863 0.950980392 0.506
#> 61  0.70098039 0.950980392 0.494
#> 62  0.78921569 0.950980392 0.752
#> 63  0.81862745 0.950980392 0.178
#> 64  0.02450980 0.931372549 0.846
#> 65  0.05392157 0.931372549 0.865
#> 66  0.08333333 0.931372549 0.045
#> 67  0.11274510 0.931372549 0.318
#> 68  0.14215686 0.931372549 0.018
#> 69  0.17156863 0.931372549 0.240
#> 70  0.20098039 0.931372549 0.705
#> 71  0.23039216 0.931372549 0.311
#> 72  0.25980392 0.931372549 0.510
#> 73  0.28921569 0.931372549 0.053
#> 74  0.31862745 0.931372549 0.564
#> 75  0.34803922 0.931372549 0.123
#> 76  0.43627451 0.931372549 0.893
#> 77  0.46568627 0.931372549 0.018
#> 78  0.49509804 0.931372549 0.783
#> 79  0.58333333 0.931372549 0.092
#> 80  0.61274510 0.931372549 0.518
#> 81  0.64215686 0.931372549 0.385
#> 82  0.73039216 0.931372549 0.072
#> 83  0.75980392 0.931372549 0.322
#> 84  0.02450980 0.911764706 0.041
#> 85  0.05392157 0.911764706 0.221
#> 86  0.08333333 0.911764706 0.811
#> 87  0.11274510 0.911764706 0.525
#> 88  0.14215686 0.911764706 0.912
#> 89  0.17156863 0.911764706 0.830
#> 90  0.20098039 0.911764706 0.049
#> 91  0.23039216 0.911764706 0.455
#> 92  0.25980392 0.911764706 0.268
#> 93  0.28921569 0.911764706 0.307
#> 94  0.31862745 0.911764706 0.506
#> 95  0.40686275 0.911764706 0.182
#> 96  0.43627451 0.911764706 0.760
#> 97  0.46568627 0.911764706 0.201
#> 98  0.55392157 0.911764706 0.260
#> 99  0.58333333 0.911764706 0.990
#> 100 0.02450980 0.852941176 0.123
#> 101 0.05392157 0.852941176 0.221
#> 102 0.08333333 0.852941176 0.912
#> 103 0.11274510 0.852941176 0.943
#> 104 0.14215686 0.852941176 0.279
#> 105 0.17156863 0.852941176 0.123
#> 106 0.20098039 0.852941176 0.795
#> 107 0.23039216 0.852941176 0.744
#> 108 0.25980392 0.852941176 0.916
#> 109 0.28921569 0.852941176 0.994
#> 110 0.31862745 0.852941176 0.939
#> 111 0.34803922 0.852941176 0.486
#> 112 0.46568627 0.852941176 0.283
#> 113 0.49509804 0.852941176 0.252
#> 114 0.61274510 0.852941176 0.502
#> 115 0.64215686 0.852941176 0.498
#> 116 0.02450980 0.833333333 0.740
#> 117 0.05392157 0.833333333 0.037
#> 118 0.08333333 0.833333333 0.564
#> 119 0.11274510 0.833333333 0.279
#> 120 0.14215686 0.833333333 0.205
#> 121 0.17156863 0.833333333 0.135
#> 122 0.20098039 0.833333333 0.326
#> 123 0.23039216 0.833333333 0.158
#> 124 0.25980392 0.833333333 0.131
#> 125 0.28921569 0.833333333 0.436
#> 126 0.40686275 0.833333333 0.041
#> 127 0.43627451 0.833333333 0.713
#> 128 0.55392157 0.833333333 0.104
#> 129 0.58333333 0.833333333 0.947
#> 130 0.02450980 0.774509804 0.318
#> 131 0.05392157 0.774509804 0.959
#> 132 0.08333333 0.774509804 0.635
#> 133 0.11274510 0.774509804 0.127
#> 134 0.14215686 0.774509804 0.424
#> 135 0.17156863 0.774509804 0.912
#> 136 0.20098039 0.774509804 0.467
#> 137 0.23039216 0.774509804 0.908
#> 138 0.25980392 0.774509804 0.596
#> 139 0.34803922 0.774509804 0.631
#> 140 0.37745098 0.774509804 0.869
#> 141 0.40686275 0.774509804 0.502
#> 142 0.49509804 0.774509804 0.982
#> 143 0.52450980 0.774509804 0.326
#> 144 0.55392157 0.774509804 0.482
#> 145 0.64215686 0.774509804 0.357
#> 146 0.67156863 0.774509804 0.627
#> 147 0.70098039 0.774509804 0.740
#> 148 0.78921569 0.774509804 0.564
#> 149 0.81862745 0.774509804 0.979
#> 150 0.02450980 0.754901961 0.584
#> 151 0.05392157 0.754901961 0.709
#> 152 0.08333333 0.754901961 0.428
#> 153 0.11274510 0.754901961 0.346
#> 154 0.14215686 0.754901961 0.760
#> 155 0.17156863 0.754901961 0.424
#> 156 0.20098039 0.754901961 0.561
#> 157 0.23039216 0.754901961 0.119
#> 158 0.25980392 0.754901961 0.303
#> 159 0.34803922 0.754901961 0.479
#> 160 0.37745098 0.754901961 0.346
#> 161 0.40686275 0.754901961 0.600
#> 162 0.49509804 0.754901961 0.076
#> 163 0.52450980 0.754901961 0.955
#> 164 0.55392157 0.754901961 0.025
#> 165 0.64215686 0.754901961 0.842
#> 166 0.67156863 0.754901961 0.631
#> 167 0.75980392 0.754901961 0.311
#> 168 0.78921569 0.754901961 0.740
#> 169 0.02450980 0.735294118 0.576
#> 170 0.05392157 0.735294118 0.439
#> 171 0.08333333 0.735294118 0.229
#> 172 0.11274510 0.735294118 0.084
#> 173 0.14215686 0.735294118 0.850
#> 174 0.17156863 0.735294118 0.236
#> 175 0.20098039 0.735294118 0.986
#> 176 0.23039216 0.735294118 0.600
#> 177 0.25980392 0.735294118 0.998
#> 178 0.28921569 0.735294118 0.377
#> 179 0.31862745 0.735294118 0.557
#> 180 0.34803922 0.735294118 0.432
#> 181 0.37745098 0.735294118 0.576
#> 182 0.40686275 0.735294118 0.432
#> 183 0.43627451 0.735294118 0.225
#> 184 0.52450980 0.735294118 0.088
#> 185 0.55392157 0.735294118 0.639
#> 186 0.58333333 0.735294118 0.432
#> 187 0.67156863 0.735294118 0.076
#> 188 0.70098039 0.735294118 0.803
#> 189 0.73039216 0.735294118 0.326
#> 190 0.75980392 0.735294118 0.756
#> 191 0.02450980 0.558823529 0.639
#> 192 0.05392157 0.558823529 0.990
#> 193 0.08333333 0.558823529 0.131
#> 194 0.11274510 0.558823529 0.881
#> 195 0.14215686 0.558823529 0.811
#> 196 0.17156863 0.558823529 0.822
#> 197 0.20098039 0.558823529 0.834
#> 198 0.23039216 0.558823529 0.732
#> 199 0.25980392 0.558823529 0.982
#> 200 0.28921569 0.558823529 0.639
#> 201 0.31862745 0.558823529 0.658
#> 202 0.34803922 0.558823529 0.529
#> 203 0.37745098 0.558823529 0.318
#> 204 0.40686275 0.558823529 0.768
#> 205 0.43627451 0.558823529 0.525
#> 206 0.46568627 0.558823529 0.732
#> 207 0.49509804 0.558823529 0.307
#> 208 0.52450980 0.558823529 0.404
#> 209 0.55392157 0.558823529 0.205
#> 210 0.58333333 0.558823529 0.982
#> 211 0.61274510 0.558823529 0.564
#> 212 0.64215686 0.558823529 0.279
#> 213 0.67156863 0.558823529 0.186
#> 214 0.70098039 0.558823529 0.756
#> 215 0.73039216 0.558823529 0.568
#> 216 0.75980392 0.558823529 0.932
#> 217 0.84803922 0.558823529 0.639
#> 218 0.87745098 0.558823529 0.701
#> 219 0.02450980 0.500000000 0.748
#> 220 0.05392157 0.500000000 0.357
#> 221 0.08333333 0.500000000 0.760
#> 222 0.11274510 0.500000000 0.377
#> 223 0.14215686 0.500000000 0.799
#> 224 0.17156863 0.500000000 0.029
#> 225 0.20098039 0.500000000 0.506
#> 226 0.23039216 0.500000000 0.818
#> 227 0.25980392 0.500000000 0.545
#> 228 0.28921569 0.500000000 0.268
#> 229 0.31862745 0.500000000 0.346
#> 230 0.34803922 0.500000000 0.369
#> 231 0.37745098 0.500000000 0.428
#> 232 0.40686275 0.500000000 0.916
#> 233 0.43627451 0.500000000 0.783
#> 234 0.46568627 0.500000000 0.736
#> 235 0.49509804 0.500000000 0.283
#> 236 0.52450980 0.500000000 0.455
#> 237 0.55392157 0.500000000 0.287
#> 238 0.58333333 0.500000000 0.697
#> 239 0.61274510 0.500000000 0.818
#> 240 0.64215686 0.500000000 0.654
#> 241 0.67156863 0.500000000 0.412
#> 242 0.70098039 0.500000000 0.951
#> 243 0.73039216 0.500000000 0.244
#> 244 0.81862745 0.500000000 0.607
#> 245 0.84803922 0.500000000 0.756
#> 246 0.02450980 0.480392157 0.689
#> 247 0.05392157 0.480392157 0.979
#> 248 0.08333333 0.480392157 0.479
#> 249 0.11274510 0.480392157 0.771
#> 250 0.14215686 0.480392157 0.572
#> 251 0.17156863 0.480392157 0.963
#> 252 0.20098039 0.480392157 0.795
#> 253 0.23039216 0.480392157 0.533
#> 254 0.25980392 0.480392157 0.596
#> 255 0.28921569 0.480392157 0.264
#> 256 0.31862745 0.480392157 0.279
#> 257 0.34803922 0.480392157 0.068
#> 258 0.37745098 0.480392157 0.564
#> 259 0.40686275 0.480392157 0.264
#> 260 0.43627451 0.480392157 0.006
#> 261 0.46568627 0.480392157 0.588
#> 262 0.49509804 0.480392157 0.521
#> 263 0.52450980 0.480392157 0.842
#> 264 0.55392157 0.480392157 0.033
#> 265 0.58333333 0.480392157 0.600
#> 266 0.61274510 0.480392157 0.268
#> 267 0.64215686 0.480392157 0.123
#> 268 0.67156863 0.480392157 0.104
#> 269 0.70098039 0.480392157 0.748
#> 270 0.78921569 0.480392157 0.018
#> 271 0.81862745 0.480392157 0.053
#> 272 0.02450980 0.460784314 0.479
#> 273 0.05392157 0.460784314 0.850
#> 274 0.08333333 0.460784314 0.424
#> 275 0.11274510 0.460784314 0.033
#> 276 0.14215686 0.460784314 0.260
#> 277 0.17156863 0.460784314 0.334
#> 278 0.20098039 0.460784314 0.135
#> 279 0.23039216 0.460784314 0.498
#> 280 0.25980392 0.460784314 0.803
#> 281 0.28921569 0.460784314 0.338
#> 282 0.31862745 0.460784314 0.510
#> 283 0.34803922 0.460784314 0.494
#> 284 0.37745098 0.460784314 0.795
#> 285 0.40686275 0.460784314 0.568
#> 286 0.43627451 0.460784314 0.107
#> 287 0.46568627 0.460784314 0.807
#> 288 0.49509804 0.460784314 0.568
#> 289 0.52450980 0.460784314 0.213
#> 290 0.55392157 0.460784314 0.748
#> 291 0.58333333 0.460784314 0.307
#> 292 0.61274510 0.460784314 0.490
#> 293 0.64215686 0.460784314 0.986
#> 294 0.67156863 0.460784314 0.424
#> 295 0.75980392 0.460784314 0.244
#> 296 0.78921569 0.460784314 0.217
#> 297 0.02450980 0.284313725 0.693
#> 298 0.05392157 0.284313725 0.115
#> 299 0.08333333 0.284313725 0.635
#> 300 0.17156863 0.284313725 0.311
#> 301 0.20098039 0.284313725 0.354
#> 302 0.28921569 0.284313725 0.979
#> 303 0.31862745 0.284313725 0.537
#> 304 0.40686275 0.284313725 0.443
#> 305 0.43627451 0.284313725 0.947
#> 306 0.52450980 0.284313725 0.451
#> 307 0.55392157 0.284313725 0.193
#> 308 0.64215686 0.284313725 0.990
#> 309 0.67156863 0.284313725 0.549
#> 310 0.75980392 0.284313725 0.768
#> 311 0.78921569 0.284313725 0.912
#> 312 0.87745098 0.284313725 0.682
#> 313 0.90686275 0.284313725 0.408
#> 314 0.02450980 0.264705882 0.408
#> 315 0.05392157 0.264705882 0.146
#> 316 0.08333333 0.264705882 0.197
#> 317 0.11274510 0.264705882 0.193
#> 318 0.14215686 0.264705882 0.408
#> 319 0.17156863 0.264705882 0.350
#> 320 0.20098039 0.264705882 0.834
#> 321 0.28921569 0.264705882 0.201
#> 322 0.31862745 0.264705882 0.861
#> 323 0.40686275 0.264705882 0.396
#> 324 0.43627451 0.264705882 0.154
#> 325 0.52450980 0.264705882 0.342
#> 326 0.55392157 0.264705882 0.369
#> 327 0.64215686 0.264705882 0.428
#> 328 0.67156863 0.264705882 0.189
#> 329 0.75980392 0.264705882 0.658
#> 330 0.78921569 0.264705882 0.920
#> 331 0.87745098 0.264705882 0.732
#> 332 0.90686275 0.264705882 0.881
#> 333 0.02450980 0.205882353 0.951
#> 334 0.05392157 0.205882353 0.197
#> 335 0.08333333 0.205882353 0.475
#> 336 0.11274510 0.205882353 0.385
#> 337 0.14215686 0.205882353 0.373
#> 338 0.17156863 0.205882353 0.029
#> 339 0.25980392 0.205882353 0.928
#> 340 0.28921569 0.205882353 0.412
#> 341 0.37745098 0.205882353 0.955
#> 342 0.40686275 0.205882353 0.271
#> 343 0.49509804 0.205882353 0.518
#> 344 0.52450980 0.205882353 0.975
#> 345 0.61274510 0.205882353 0.369
#> 346 0.64215686 0.205882353 0.311
#> 347 0.73039216 0.205882353 0.037
#> 348 0.75980392 0.205882353 0.666
#> 349 0.84803922 0.205882353 0.920
#> 350 0.87745098 0.205882353 0.045
#> 351 0.02450980 0.186274510 0.201
#> 352 0.05392157 0.186274510 0.744
#> 353 0.08333333 0.186274510 0.131
#> 354 0.11274510 0.186274510 0.709
#> 355 0.14215686 0.186274510 0.998
#> 356 0.17156863 0.186274510 0.943
#> 357 0.25980392 0.186274510 0.592
#> 358 0.28921569 0.186274510 0.729
#> 359 0.37745098 0.186274510 0.486
#> 360 0.40686275 0.186274510 0.768
#> 361 0.49509804 0.186274510 0.006
#> 362 0.52450980 0.186274510 0.557
#> 363 0.61274510 0.186274510 0.459
#> 364 0.64215686 0.186274510 0.330
#> 365 0.73039216 0.186274510 0.834
#> 366 0.75980392 0.186274510 0.975
#> 367 0.84803922 0.186274510 0.658
#> 368 0.02450980 0.166666667 0.236
#> 369 0.05392157 0.166666667 0.818
#> 370 0.08333333 0.166666667 0.725
#> 371 0.11274510 0.166666667 0.975
#> 372 0.14215686 0.166666667 0.268
#> 373 0.17156863 0.166666667 0.877
#> 374 0.20098039 0.166666667 0.486
#> 375 0.23039216 0.166666667 0.307
#> 376 0.25980392 0.166666667 0.396
#> 377 0.28921569 0.166666667 0.760
#> 378 0.37745098 0.166666667 0.104
#> 379 0.40686275 0.166666667 0.420
#> 380 0.49509804 0.166666667 0.658
#> 381 0.52450980 0.166666667 0.295
#> 382 0.61274510 0.166666667 0.209
#> 383 0.64215686 0.166666667 0.006
#> 384 0.73039216 0.166666667 0.107
#> 385 0.75980392 0.166666667 0.217
#> 386 0.78921569 0.166666667 0.115
#> 387 0.81862745 0.166666667 0.686
#> 388 0.02450980 0.107843137 0.193
#> 389 0.05392157 0.107843137 0.982
#> 390 0.08333333 0.107843137 0.943
#> 391 0.11274510 0.107843137 0.689
#> 392 0.14215686 0.107843137 0.740
#> 393 0.17156863 0.107843137 0.732
#> 394 0.20098039 0.107843137 0.650
#> 395 0.28921569 0.107843137 0.170
#> 396 0.31862745 0.107843137 0.920
#> 397 0.40686275 0.107843137 0.564
#> 398 0.43627451 0.107843137 0.537
#> 399 0.52450980 0.107843137 0.021
#> 400 0.55392157 0.107843137 0.365
#> 401 0.64215686 0.107843137 0.686
#> 402 0.67156863 0.107843137 0.416
#> 403 0.75980392 0.107843137 0.756
#> 404 0.78921569 0.107843137 0.775
#> 405 0.87745098 0.107843137 0.572
#> 406 0.90686275 0.107843137 0.189
#> 407 0.02450980 0.049019608 0.756
#> 408 0.05392157 0.049019608 0.096
#> 409 0.08333333 0.049019608 0.623
#> 410 0.11274510 0.049019608 0.416
#> 411 0.14215686 0.049019608 0.775
#> 412 0.17156863 0.049019608 0.807
#> 413 0.20098039 0.049019608 0.967
#> 414 0.23039216 0.049019608 0.217
#> 415 0.25980392 0.049019608 0.865
#> 416 0.28921569 0.049019608 0.416
#> 417 0.31862745 0.049019608 0.514
#> 418 0.40686275 0.049019608 0.775
#> 419 0.43627451 0.049019608 0.135
#> 420 0.52450980 0.049019608 0.412
#> 421 0.55392157 0.049019608 0.662
#> 422 0.64215686 0.049019608 0.904
#> 423 0.67156863 0.049019608 0.346
#> 424 0.75980392 0.049019608 0.107
#> 425 0.78921569 0.049019608 0.932
#> 426 0.87745098 0.049019608 0.201
#> 427 0.90686275 0.049019608 0.057
#> 428 0.02450980 0.029411765 0.631
#> 429 0.05392157 0.029411765 0.611
#> 430 0.08333333 0.029411765 0.037
#> 431 0.11274510 0.029411765 0.256
#> 432 0.14215686 0.029411765 0.775
#> 433 0.17156863 0.029411765 0.998
#> 434 0.20098039 0.029411765 0.959
#> 435 0.23039216 0.029411765 0.748
#> 436 0.31862745 0.029411765 0.682
#> 437 0.34803922 0.029411765 0.248
#> 438 0.43627451 0.029411765 0.607
#> 439 0.46568627 0.029411765 0.572
#> 440 0.55392157 0.029411765 0.053
#> 441 0.58333333 0.029411765 0.373
#> 442 0.67156863 0.029411765 0.893
#> 443 0.70098039 0.029411765 0.393
#> 444 0.78921569 0.029411765 0.518
#> 445 0.81862745 0.029411765 0.178
#> 446 0.84803922 0.029411765 0.193
#> 447 0.02450980 0.009803922 0.428
#> 448 0.05392157 0.009803922 0.338
#> 449 0.08333333 0.009803922 0.275
#> 450 0.11274510 0.009803922 0.662
#> 451 0.14215686 0.009803922 0.783
#> 452 0.17156863 0.009803922 0.439
#> 453 0.20098039 0.009803922 0.928
#> 454 0.23039216 0.009803922 0.271
#> 455 0.25980392 0.009803922 0.658
#> 456 0.28921569 0.009803922 0.381
#> 457 0.31862745 0.009803922 0.936
#> 458 0.34803922 0.009803922 0.736
#> 459 0.43627451 0.009803922 0.588
#> 460 0.46568627 0.009803922 0.814
#> 461 0.55392157 0.009803922 0.881
#> 462 0.58333333 0.009803922 0.744
#> 463 0.67156863 0.009803922 0.502
#> 464 0.70098039 0.009803922 0.986
#> 465 0.73039216 0.009803922 0.557
#> 466 0.75980392 0.009803922 0.877

## Default (i.e. scaled to 0-1, 2 dp)
rasterise_probabilities(
    d$locations,
    d$probabilities,
    d$sequences
)
#>              x           y value
#> 1   0.02450980 0.990196078  0.12
#> 2   0.05392157 0.990196078  0.62
#> 3   0.08333333 0.990196078  0.61
#> 4   0.11274510 0.990196078  0.62
#> 5   0.14215686 0.990196078  0.86
#> 6   0.17156863 0.990196078  0.64
#> 7   0.20098039 0.990196078  0.01
#> 8   0.23039216 0.990196078  0.23
#> 9   0.25980392 0.990196078  0.67
#> 10  0.34803922 0.990196078  0.51
#> 11  0.37745098 0.990196078  0.69
#> 12  0.40686275 0.990196078  0.54
#> 13  0.49509804 0.990196078  0.28
#> 14  0.52450980 0.990196078  0.92
#> 15  0.55392157 0.990196078  0.29
#> 16  0.64215686 0.990196078  0.84
#> 17  0.67156863 0.990196078  0.29
#> 18  0.70098039 0.990196078  0.27
#> 19  0.78921569 0.990196078  0.19
#> 20  0.81862745 0.990196078  0.23
#> 21  0.84803922 0.990196078  0.32
#> 22  0.93627451 0.990196078  0.30
#> 23  0.96568627 0.990196078  0.16
#> 24  0.02450980 0.970588235  0.67
#> 25  0.05392157 0.970588235  0.92
#> 26  0.08333333 0.970588235  0.47
#> 27  0.11274510 0.970588235  0.14
#> 28  0.14215686 0.970588235  0.54
#> 29  0.17156863 0.970588235  0.20
#> 30  0.25980392 0.970588235  0.90
#> 31  0.28921569 0.970588235  0.39
#> 32  0.31862745 0.970588235  0.31
#> 33  0.40686275 0.970588235  0.16
#> 34  0.43627451 0.970588235  0.90
#> 35  0.46568627 0.970588235  0.17
#> 36  0.55392157 0.970588235  0.90
#> 37  0.58333333 0.970588235  0.13
#> 38  0.61274510 0.970588235  0.13
#> 39  0.70098039 0.970588235  0.11
#> 40  0.73039216 0.970588235  0.51
#> 41  0.75980392 0.970588235  0.30
#> 42  0.84803922 0.970588235  0.03
#> 43  0.87745098 0.970588235  0.31
#> 44  0.02450980 0.950980392  0.81
#> 45  0.05392157 0.950980392  0.55
#> 46  0.08333333 0.950980392  0.65
#> 47  0.11274510 0.950980392  0.31
#> 48  0.14215686 0.950980392  0.62
#> 49  0.17156863 0.950980392  0.33
#> 50  0.20098039 0.950980392  0.50
#> 51  0.23039216 0.950980392  0.68
#> 52  0.25980392 0.950980392  0.49
#> 53  0.34803922 0.950980392  0.24
#> 54  0.37745098 0.950980392  0.76
#> 55  0.40686275 0.950980392  0.08
#> 56  0.49509804 0.950980392  0.31
#> 57  0.52450980 0.950980392  0.72
#> 58  0.55392157 0.950980392  0.51
#> 59  0.64215686 0.950980392  0.15
#> 60  0.67156863 0.950980392  0.51
#> 61  0.70098039 0.950980392  0.49
#> 62  0.78921569 0.950980392  0.75
#> 63  0.81862745 0.950980392  0.18
#> 64  0.02450980 0.931372549  0.85
#> 65  0.05392157 0.931372549  0.87
#> 66  0.08333333 0.931372549  0.04
#> 67  0.11274510 0.931372549  0.32
#> 68  0.14215686 0.931372549  0.02
#> 69  0.17156863 0.931372549  0.24
#> 70  0.20098039 0.931372549  0.71
#> 71  0.23039216 0.931372549  0.31
#> 72  0.25980392 0.931372549  0.51
#> 73  0.28921569 0.931372549  0.05
#> 74  0.31862745 0.931372549  0.56
#> 75  0.34803922 0.931372549  0.12
#> 76  0.43627451 0.931372549  0.89
#> 77  0.46568627 0.931372549  0.02
#> 78  0.49509804 0.931372549  0.78
#> 79  0.58333333 0.931372549  0.09
#> 80  0.61274510 0.931372549  0.52
#> 81  0.64215686 0.931372549  0.38
#> 82  0.73039216 0.931372549  0.07
#> 83  0.75980392 0.931372549  0.32
#> 84  0.02450980 0.911764706  0.04
#> 85  0.05392157 0.911764706  0.22
#> 86  0.08333333 0.911764706  0.81
#> 87  0.11274510 0.911764706  0.53
#> 88  0.14215686 0.911764706  0.91
#> 89  0.17156863 0.911764706  0.83
#> 90  0.20098039 0.911764706  0.05
#> 91  0.23039216 0.911764706  0.46
#> 92  0.25980392 0.911764706  0.27
#> 93  0.28921569 0.911764706  0.31
#> 94  0.31862745 0.911764706  0.51
#> 95  0.40686275 0.911764706  0.18
#> 96  0.43627451 0.911764706  0.76
#> 97  0.46568627 0.911764706  0.20
#> 98  0.55392157 0.911764706  0.26
#> 99  0.58333333 0.911764706  0.99
#> 100 0.02450980 0.852941176  0.12
#> 101 0.05392157 0.852941176  0.22
#> 102 0.08333333 0.852941176  0.91
#> 103 0.11274510 0.852941176  0.94
#> 104 0.14215686 0.852941176  0.28
#> 105 0.17156863 0.852941176  0.12
#> 106 0.20098039 0.852941176  0.79
#> 107 0.23039216 0.852941176  0.74
#> 108 0.25980392 0.852941176  0.92
#> 109 0.28921569 0.852941176  0.99
#> 110 0.31862745 0.852941176  0.94
#> 111 0.34803922 0.852941176  0.49
#> 112 0.46568627 0.852941176  0.28
#> 113 0.49509804 0.852941176  0.25
#> 114 0.61274510 0.852941176  0.50
#> 115 0.64215686 0.852941176  0.50
#> 116 0.02450980 0.833333333  0.74
#> 117 0.05392157 0.833333333  0.04
#> 118 0.08333333 0.833333333  0.56
#> 119 0.11274510 0.833333333  0.28
#> 120 0.14215686 0.833333333  0.21
#> 121 0.17156863 0.833333333  0.13
#> 122 0.20098039 0.833333333  0.33
#> 123 0.23039216 0.833333333  0.16
#> 124 0.25980392 0.833333333  0.13
#> 125 0.28921569 0.833333333  0.44
#> 126 0.40686275 0.833333333  0.04
#> 127 0.43627451 0.833333333  0.71
#> 128 0.55392157 0.833333333  0.10
#> 129 0.58333333 0.833333333  0.95
#> 130 0.02450980 0.774509804  0.32
#> 131 0.05392157 0.774509804  0.96
#> 132 0.08333333 0.774509804  0.63
#> 133 0.11274510 0.774509804  0.13
#> 134 0.14215686 0.774509804  0.42
#> 135 0.17156863 0.774509804  0.91
#> 136 0.20098039 0.774509804  0.47
#> 137 0.23039216 0.774509804  0.91
#> 138 0.25980392 0.774509804  0.60
#> 139 0.34803922 0.774509804  0.63
#> 140 0.37745098 0.774509804  0.87
#> 141 0.40686275 0.774509804  0.50
#> 142 0.49509804 0.774509804  0.98
#> 143 0.52450980 0.774509804  0.33
#> 144 0.55392157 0.774509804  0.48
#> 145 0.64215686 0.774509804  0.36
#> 146 0.67156863 0.774509804  0.63
#> 147 0.70098039 0.774509804  0.74
#> 148 0.78921569 0.774509804  0.56
#> 149 0.81862745 0.774509804  0.98
#> 150 0.02450980 0.754901961  0.58
#> 151 0.05392157 0.754901961  0.71
#> 152 0.08333333 0.754901961  0.43
#> 153 0.11274510 0.754901961  0.35
#> 154 0.14215686 0.754901961  0.76
#> 155 0.17156863 0.754901961  0.42
#> 156 0.20098039 0.754901961  0.56
#> 157 0.23039216 0.754901961  0.12
#> 158 0.25980392 0.754901961  0.30
#> 159 0.34803922 0.754901961  0.48
#> 160 0.37745098 0.754901961  0.35
#> 161 0.40686275 0.754901961  0.60
#> 162 0.49509804 0.754901961  0.08
#> 163 0.52450980 0.754901961  0.96
#> 164 0.55392157 0.754901961  0.03
#> 165 0.64215686 0.754901961  0.84
#> 166 0.67156863 0.754901961  0.63
#> 167 0.75980392 0.754901961  0.31
#> 168 0.78921569 0.754901961  0.74
#> 169 0.02450980 0.735294118  0.58
#> 170 0.05392157 0.735294118  0.44
#> 171 0.08333333 0.735294118  0.23
#> 172 0.11274510 0.735294118  0.08
#> 173 0.14215686 0.735294118  0.85
#> 174 0.17156863 0.735294118  0.24
#> 175 0.20098039 0.735294118  0.99
#> 176 0.23039216 0.735294118  0.60
#> 177 0.25980392 0.735294118  1.00
#> 178 0.28921569 0.735294118  0.38
#> 179 0.31862745 0.735294118  0.56
#> 180 0.34803922 0.735294118  0.43
#> 181 0.37745098 0.735294118  0.58
#> 182 0.40686275 0.735294118  0.43
#> 183 0.43627451 0.735294118  0.22
#> 184 0.52450980 0.735294118  0.09
#> 185 0.55392157 0.735294118  0.64
#> 186 0.58333333 0.735294118  0.43
#> 187 0.67156863 0.735294118  0.08
#> 188 0.70098039 0.735294118  0.80
#> 189 0.73039216 0.735294118  0.33
#> 190 0.75980392 0.735294118  0.76
#> 191 0.02450980 0.558823529  0.64
#> 192 0.05392157 0.558823529  0.99
#> 193 0.08333333 0.558823529  0.13
#> 194 0.11274510 0.558823529  0.88
#> 195 0.14215686 0.558823529  0.81
#> 196 0.17156863 0.558823529  0.82
#> 197 0.20098039 0.558823529  0.83
#> 198 0.23039216 0.558823529  0.73
#> 199 0.25980392 0.558823529  0.98
#> 200 0.28921569 0.558823529  0.64
#> 201 0.31862745 0.558823529  0.66
#> 202 0.34803922 0.558823529  0.53
#> 203 0.37745098 0.558823529  0.32
#> 204 0.40686275 0.558823529  0.77
#> 205 0.43627451 0.558823529  0.53
#> 206 0.46568627 0.558823529  0.73
#> 207 0.49509804 0.558823529  0.31
#> 208 0.52450980 0.558823529  0.40
#> 209 0.55392157 0.558823529  0.21
#> 210 0.58333333 0.558823529  0.98
#> 211 0.61274510 0.558823529  0.56
#> 212 0.64215686 0.558823529  0.28
#> 213 0.67156863 0.558823529  0.19
#> 214 0.70098039 0.558823529  0.76
#> 215 0.73039216 0.558823529  0.57
#> 216 0.75980392 0.558823529  0.93
#> 217 0.84803922 0.558823529  0.64
#> 218 0.87745098 0.558823529  0.70
#> 219 0.02450980 0.500000000  0.75
#> 220 0.05392157 0.500000000  0.36
#> 221 0.08333333 0.500000000  0.76
#> 222 0.11274510 0.500000000  0.38
#> 223 0.14215686 0.500000000  0.80
#> 224 0.17156863 0.500000000  0.03
#> 225 0.20098039 0.500000000  0.51
#> 226 0.23039216 0.500000000  0.82
#> 227 0.25980392 0.500000000  0.54
#> 228 0.28921569 0.500000000  0.27
#> 229 0.31862745 0.500000000  0.35
#> 230 0.34803922 0.500000000  0.37
#> 231 0.37745098 0.500000000  0.43
#> 232 0.40686275 0.500000000  0.92
#> 233 0.43627451 0.500000000  0.78
#> 234 0.46568627 0.500000000  0.74
#> 235 0.49509804 0.500000000  0.28
#> 236 0.52450980 0.500000000  0.46
#> 237 0.55392157 0.500000000  0.29
#> 238 0.58333333 0.500000000  0.70
#> 239 0.61274510 0.500000000  0.82
#> 240 0.64215686 0.500000000  0.65
#> 241 0.67156863 0.500000000  0.41
#> 242 0.70098039 0.500000000  0.95
#> 243 0.73039216 0.500000000  0.24
#> 244 0.81862745 0.500000000  0.61
#> 245 0.84803922 0.500000000  0.76
#> 246 0.02450980 0.480392157  0.69
#> 247 0.05392157 0.480392157  0.98
#> 248 0.08333333 0.480392157  0.48
#> 249 0.11274510 0.480392157  0.77
#> 250 0.14215686 0.480392157  0.57
#> 251 0.17156863 0.480392157  0.96
#> 252 0.20098039 0.480392157  0.79
#> 253 0.23039216 0.480392157  0.53
#> 254 0.25980392 0.480392157  0.60
#> 255 0.28921569 0.480392157  0.26
#> 256 0.31862745 0.480392157  0.28
#> 257 0.34803922 0.480392157  0.07
#> 258 0.37745098 0.480392157  0.56
#> 259 0.40686275 0.480392157  0.26
#> 260 0.43627451 0.480392157  0.01
#> 261 0.46568627 0.480392157  0.59
#> 262 0.49509804 0.480392157  0.52
#> 263 0.52450980 0.480392157  0.84
#> 264 0.55392157 0.480392157  0.03
#> 265 0.58333333 0.480392157  0.60
#> 266 0.61274510 0.480392157  0.27
#> 267 0.64215686 0.480392157  0.12
#> 268 0.67156863 0.480392157  0.10
#> 269 0.70098039 0.480392157  0.75
#> 270 0.78921569 0.480392157  0.02
#> 271 0.81862745 0.480392157  0.05
#> 272 0.02450980 0.460784314  0.48
#> 273 0.05392157 0.460784314  0.85
#> 274 0.08333333 0.460784314  0.42
#> 275 0.11274510 0.460784314  0.03
#> 276 0.14215686 0.460784314  0.26
#> 277 0.17156863 0.460784314  0.33
#> 278 0.20098039 0.460784314  0.13
#> 279 0.23039216 0.460784314  0.50
#> 280 0.25980392 0.460784314  0.80
#> 281 0.28921569 0.460784314  0.34
#> 282 0.31862745 0.460784314  0.51
#> 283 0.34803922 0.460784314  0.49
#> 284 0.37745098 0.460784314  0.79
#> 285 0.40686275 0.460784314  0.57
#> 286 0.43627451 0.460784314  0.11
#> 287 0.46568627 0.460784314  0.81
#> 288 0.49509804 0.460784314  0.57
#> 289 0.52450980 0.460784314  0.21
#> 290 0.55392157 0.460784314  0.75
#> 291 0.58333333 0.460784314  0.31
#> 292 0.61274510 0.460784314  0.49
#> 293 0.64215686 0.460784314  0.99
#> 294 0.67156863 0.460784314  0.42
#> 295 0.75980392 0.460784314  0.24
#> 296 0.78921569 0.460784314  0.22
#> 297 0.02450980 0.284313725  0.69
#> 298 0.05392157 0.284313725  0.12
#> 299 0.08333333 0.284313725  0.63
#> 300 0.17156863 0.284313725  0.31
#> 301 0.20098039 0.284313725  0.35
#> 302 0.28921569 0.284313725  0.98
#> 303 0.31862745 0.284313725  0.54
#> 304 0.40686275 0.284313725  0.44
#> 305 0.43627451 0.284313725  0.95
#> 306 0.52450980 0.284313725  0.45
#> 307 0.55392157 0.284313725  0.19
#> 308 0.64215686 0.284313725  0.99
#> 309 0.67156863 0.284313725  0.55
#> 310 0.75980392 0.284313725  0.77
#> 311 0.78921569 0.284313725  0.91
#> 312 0.87745098 0.284313725  0.68
#> 313 0.90686275 0.284313725  0.41
#> 314 0.02450980 0.264705882  0.41
#> 315 0.05392157 0.264705882  0.15
#> 316 0.08333333 0.264705882  0.20
#> 317 0.11274510 0.264705882  0.19
#> 318 0.14215686 0.264705882  0.41
#> 319 0.17156863 0.264705882  0.35
#> 320 0.20098039 0.264705882  0.83
#> 321 0.28921569 0.264705882  0.20
#> 322 0.31862745 0.264705882  0.86
#> 323 0.40686275 0.264705882  0.40
#> 324 0.43627451 0.264705882  0.15
#> 325 0.52450980 0.264705882  0.34
#> 326 0.55392157 0.264705882  0.37
#> 327 0.64215686 0.264705882  0.43
#> 328 0.67156863 0.264705882  0.19
#> 329 0.75980392 0.264705882  0.66
#> 330 0.78921569 0.264705882  0.92
#> 331 0.87745098 0.264705882  0.73
#> 332 0.90686275 0.264705882  0.88
#> 333 0.02450980 0.205882353  0.95
#> 334 0.05392157 0.205882353  0.20
#> 335 0.08333333 0.205882353  0.47
#> 336 0.11274510 0.205882353  0.38
#> 337 0.14215686 0.205882353  0.37
#> 338 0.17156863 0.205882353  0.03
#> 339 0.25980392 0.205882353  0.93
#> 340 0.28921569 0.205882353  0.41
#> 341 0.37745098 0.205882353  0.96
#> 342 0.40686275 0.205882353  0.27
#> 343 0.49509804 0.205882353  0.52
#> 344 0.52450980 0.205882353  0.97
#> 345 0.61274510 0.205882353  0.37
#> 346 0.64215686 0.205882353  0.31
#> 347 0.73039216 0.205882353  0.04
#> 348 0.75980392 0.205882353  0.67
#> 349 0.84803922 0.205882353  0.92
#> 350 0.87745098 0.205882353  0.04
#> 351 0.02450980 0.186274510  0.20
#> 352 0.05392157 0.186274510  0.74
#> 353 0.08333333 0.186274510  0.13
#> 354 0.11274510 0.186274510  0.71
#> 355 0.14215686 0.186274510  1.00
#> 356 0.17156863 0.186274510  0.94
#> 357 0.25980392 0.186274510  0.59
#> 358 0.28921569 0.186274510  0.73
#> 359 0.37745098 0.186274510  0.49
#> 360 0.40686275 0.186274510  0.77
#> 361 0.49509804 0.186274510  0.01
#> 362 0.52450980 0.186274510  0.56
#> 363 0.61274510 0.186274510  0.46
#> 364 0.64215686 0.186274510  0.33
#> 365 0.73039216 0.186274510  0.83
#> 366 0.75980392 0.186274510  0.97
#> 367 0.84803922 0.186274510  0.66
#> 368 0.02450980 0.166666667  0.24
#> 369 0.05392157 0.166666667  0.82
#> 370 0.08333333 0.166666667  0.72
#> 371 0.11274510 0.166666667  0.97
#> 372 0.14215686 0.166666667  0.27
#> 373 0.17156863 0.166666667  0.88
#> 374 0.20098039 0.166666667  0.49
#> 375 0.23039216 0.166666667  0.31
#> 376 0.25980392 0.166666667  0.40
#> 377 0.28921569 0.166666667  0.76
#> 378 0.37745098 0.166666667  0.10
#> 379 0.40686275 0.166666667  0.42
#> 380 0.49509804 0.166666667  0.66
#> 381 0.52450980 0.166666667  0.29
#> 382 0.61274510 0.166666667  0.21
#> 383 0.64215686 0.166666667  0.01
#> 384 0.73039216 0.166666667  0.11
#> 385 0.75980392 0.166666667  0.22
#> 386 0.78921569 0.166666667  0.12
#> 387 0.81862745 0.166666667  0.69
#> 388 0.02450980 0.107843137  0.19
#> 389 0.05392157 0.107843137  0.98
#> 390 0.08333333 0.107843137  0.94
#> 391 0.11274510 0.107843137  0.69
#> 392 0.14215686 0.107843137  0.74
#> 393 0.17156863 0.107843137  0.73
#> 394 0.20098039 0.107843137  0.65
#> 395 0.28921569 0.107843137  0.17
#> 396 0.31862745 0.107843137  0.92
#> 397 0.40686275 0.107843137  0.56
#> 398 0.43627451 0.107843137  0.54
#> 399 0.52450980 0.107843137  0.02
#> 400 0.55392157 0.107843137  0.37
#> 401 0.64215686 0.107843137  0.69
#> 402 0.67156863 0.107843137  0.42
#> 403 0.75980392 0.107843137  0.76
#> 404 0.78921569 0.107843137  0.78
#> 405 0.87745098 0.107843137  0.57
#> 406 0.90686275 0.107843137  0.19
#> 407 0.02450980 0.049019608  0.76
#> 408 0.05392157 0.049019608  0.10
#> 409 0.08333333 0.049019608  0.62
#> 410 0.11274510 0.049019608  0.42
#> 411 0.14215686 0.049019608  0.78
#> 412 0.17156863 0.049019608  0.81
#> 413 0.20098039 0.049019608  0.97
#> 414 0.23039216 0.049019608  0.22
#> 415 0.25980392 0.049019608  0.87
#> 416 0.28921569 0.049019608  0.42
#> 417 0.31862745 0.049019608  0.51
#> 418 0.40686275 0.049019608  0.78
#> 419 0.43627451 0.049019608  0.13
#> 420 0.52450980 0.049019608  0.41
#> 421 0.55392157 0.049019608  0.66
#> 422 0.64215686 0.049019608  0.90
#> 423 0.67156863 0.049019608  0.35
#> 424 0.75980392 0.049019608  0.11
#> 425 0.78921569 0.049019608  0.93
#> 426 0.87745098 0.049019608  0.20
#> 427 0.90686275 0.049019608  0.06
#> 428 0.02450980 0.029411765  0.63
#> 429 0.05392157 0.029411765  0.61
#> 430 0.08333333 0.029411765  0.04
#> 431 0.11274510 0.029411765  0.26
#> 432 0.14215686 0.029411765  0.78
#> 433 0.17156863 0.029411765  1.00
#> 434 0.20098039 0.029411765  0.96
#> 435 0.23039216 0.029411765  0.75
#> 436 0.31862745 0.029411765  0.68
#> 437 0.34803922 0.029411765  0.25
#> 438 0.43627451 0.029411765  0.61
#> 439 0.46568627 0.029411765  0.57
#> 440 0.55392157 0.029411765  0.05
#> 441 0.58333333 0.029411765  0.37
#> 442 0.67156863 0.029411765  0.89
#> 443 0.70098039 0.029411765  0.39
#> 444 0.78921569 0.029411765  0.52
#> 445 0.81862745 0.029411765  0.18
#> 446 0.84803922 0.029411765  0.19
#> 447 0.02450980 0.009803922  0.43
#> 448 0.05392157 0.009803922  0.34
#> 449 0.08333333 0.009803922  0.28
#> 450 0.11274510 0.009803922  0.66
#> 451 0.14215686 0.009803922  0.78
#> 452 0.17156863 0.009803922  0.44
#> 453 0.20098039 0.009803922  0.93
#> 454 0.23039216 0.009803922  0.27
#> 455 0.25980392 0.009803922  0.66
#> 456 0.28921569 0.009803922  0.38
#> 457 0.31862745 0.009803922  0.94
#> 458 0.34803922 0.009803922  0.74
#> 459 0.43627451 0.009803922  0.59
#> 460 0.46568627 0.009803922  0.81
#> 461 0.55392157 0.009803922  0.88
#> 462 0.58333333 0.009803922  0.74
#> 463 0.67156863 0.009803922  0.50
#> 464 0.70098039 0.009803922  0.99
#> 465 0.73039216 0.009803922  0.56
#> 466 0.75980392 0.009803922  0.88