pkgdown/google0acd00d8574ec2df.html

ggDNAvis: 'ggplot2'-Based Tools for Visualising DNA Sequences and Modifications

Uses 'ggplot2' to visualise either (a) a single DNA/RNA sequence split across multiple lines, (b) multiple DNA/RNA sequences, each occupying a whole line, or (c) base modifications such as DNA methylation called by modified bases models in Dorado or Guppy. Functions starting with visualise_<something>() are the main plotting functions, and functions starting with extract_<something>() are key helper functions for reading files and reformatting data. Source code is available at https://github.com/ejade42/ggDNAvis and a full non-expert user guide is available at https://ejade42.github.io/ggDNAvis/.

See also

Author

Maintainer: Evelyn Jade evelynjade42@gmail.com (ORCID) [copyright holder]